requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.7
loguru==0.7.2
requests==2.32.3
urllib3==2.2.2
win32-setctime==1.1.0
asyncmock==0.4.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings=2.7.0